Fix type for odometer in PlateHistory

+++ Logistica-API/Logistica-Data/DataModels/SatControlModels/PlateHistory.cs
@@ -75,7 +75,7 @@ namespace Logistica_Data.DataModels.SatControlModels
         public string? LastOff { get; set; }
 
         [XmlElement(ElementName = "Odometer")]
-        public int? Odometer { get; set; }
+        public double? Odometer { get; set; }
 
     }
